Split up code into seperate directories
