+++ CMakeLists.txt
@@ -4,16 +4,17 @@ project(x69Emulator)
 
 include(CTest)
 
-set(emu_sources 
-	"source/Emulator.h"
-	"source/Emulator.cpp"
-)
-
-add_library(${PROJECT_NAME} STATIC "include/${PROJECT_NAME}.h" "source/${PROJECT_NAME}.cpp" ${emu_sources})
+add_library(${PROJECT_NAME} STATIC "include/${PROJECT_NAME}.h" "source/${PROJECT_NAME}.cpp")
 target_include_directories(${PROJECT_NAME} PUBLIC "include" PRIVATE "source")
 set_target_properties(${PROJECT_NAME} PROPERTIES CXX_STANDARD 20 CXX_STANDARD_REQUIRED True)
 
+
+add_subdirectory("source")
+
+
 add_subdirectory("peripherals")
+
+
 
 ###
 ### TEMPORARY
