+++ source/CPU/CMakeLists.txt
@@ -1,0 +1,6 @@
+target_sources(${PROJECT_NAME} PRIVATE
+
+	"${CMAKE_CURRENT_SOURCE_DIR}/CPU.h"
+	"${CMAKE_CURRENT_SOURCE_DIR}/CPU.cpp"
+
+)
