+++ source/Instruction/Instruction.cpp
@@ -1,0 +1,6 @@
+#include "Instruction.h"
+
+namespace x69::emu
+{
+
+};
