+++ source/Peripheral/CMakeLists.txt
@@ -1,0 +1,6 @@
+target_sources(${PROJECT_NAME} PRIVATE
+
+	"${CMAKE_CURRENT_SOURCE_DIR}/Peripheral.h"
+	"${CMAKE_CURRENT_SOURCE_DIR}/Peripheral.cpp"
+
+)
