+++ source/Register/Register.cpp
@@ -1,0 +1,6 @@
+#include "Register.h"
+
+namespace x69::emu
+{
+
+};
