+++ source/Spec/Spec.cpp
@@ -1,0 +1,6 @@
+#include "Spec.h"
+
+namespace x69::emu
+{
+
+};
